third_party/WebKit/LayoutTests/external/wpt/wasm/idlharness.any-expected.txt

This is a testharness.js-based test.
Found 93 tests; 19 PASS, 74 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS wasm-js-api interfaces.
FAIL WebAssembly namespace: operation validate(BufferSource) assert_true: property should be enumerable expected true got false
FAIL WebAssembly namespace: operation compile(BufferSource) assert_true: property should be enumerable expected true got false
FAIL WebAssembly namespace: operation instantiate(BufferSource, object) assert_true: property should be enumerable expected true got false
FAIL WebAssembly namespace: operation instantiate(Module, object) assert_true: property should be enumerable expected true got false
FAIL Module interface: existence and properties of interface object assert_own_property: self does not have own property "Module" expected property "Module" missing
FAIL Module interface object length assert_own_property: self does not have own property "Module" expected property "Module" missing
FAIL Module interface object name assert_own_property: self does not have own property "Module" expected property "Module" missing
FAIL Module interface: existence and properties of interface prototype object assert_own_property: self does not have own property "Module" expected property "Module" missing
FAIL Module interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "Module" expected property "Module" missing
FAIL Module interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "Module" expected property "Module" missing
FAIL Module interface: operation exports(Module) assert_own_property: self does not have own property "Module" expected property "Module" missing
PASS Unscopable handled correctly for exports(Module) on Module
FAIL Module interface: operation imports(Module) assert_own_property: self does not have own property "Module" expected property "Module" missing
PASS Unscopable handled correctly for imports(Module) on Module
FAIL Module interface: operation customSections(Module, USVString) assert_own_property: self does not have own property "Module" expected property "Module" missing
PASS Unscopable handled correctly for customSections(Module, USVString) on Module
FAIL Module must be primary interface of mod assert_own_property: self does not have own property "Module" expected property "Module" missing
FAIL Stringification of mod assert_equals: class string of mod expected "[object Module]" but got "[object WebAssembly.Module]"
PASS Module interface: mod must inherit property "exports(Module)" with the proper type
PASS Module interface: calling exports(Module) on mod with too few arguments must throw TypeError
PASS Module interface: mod must inherit property "imports(Module)" with the proper type
PASS Module interface: calling imports(Module) on mod with too few arguments must throw TypeError
PASS Module interface: mod must inherit property "customSections(Module, USVString)" with the proper type
PASS Module interface: calling customSections(Module, USVString) on mod with too few arguments must throw TypeError
FAIL Instance interface: existence and properties of interface object assert_own_property: self does not have own property "Instance" expected property "Instance" missing
FAIL Instance interface object length assert_own_property: self does not have own property "Instance" expected property "Instance" missing
FAIL Instance interface object name assert_own_property: self does not have own property "Instance" expected property "Instance" missing
FAIL Instance interface: existence and properties of interface prototype object assert_own_property: self does not have own property "Instance" expected property "Instance" missing
FAIL Instance interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "Instance" expected property "Instance" missing
FAIL Instance interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "Instance" expected property "Instance" missing
FAIL Instance interface: attribute exports assert_own_property: self does not have own property "Instance" expected property "Instance" missing
PASS Unscopable handled correctly for exports property on Instance
FAIL Instance must be primary interface of instance assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: instance is not defined"
FAIL Stringification of instance assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: instance is not defined"
FAIL Instance interface: instance must inherit property "exports" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: instance is not defined"
FAIL Memory interface: existence and properties of interface object assert_own_property: self does not have own property "Memory" expected property "Memory" missing
FAIL Memory interface object length assert_own_property: self does not have own property "Memory" expected property "Memory" missing
FAIL Memory interface object name assert_own_property: self does not have own property "Memory" expected property "Memory" missing
FAIL Memory interface: existence and properties of interface prototype object assert_own_property: self does not have own property "Memory" expected property "Memory" missing
FAIL Memory interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "Memory" expected property "Memory" missing
FAIL Memory interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "Memory" expected property "Memory" missing
FAIL Memory interface: operation grow(unsigned long) assert_own_property: self does not have own property "Memory" expected property "Memory" missing
PASS Unscopable handled correctly for grow(unsigned long) on Memory
FAIL Memory interface: attribute buffer assert_own_property: self does not have own property "Memory" expected property "Memory" missing
PASS Unscopable handled correctly for buffer property on Memory
FAIL Memory must be primary interface of memory assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: memory is not defined"
FAIL Stringification of memory assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: memory is not defined"
FAIL Memory interface: memory must inherit property "grow(unsigned long)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: memory is not defined"
FAIL Memory interface: calling grow(unsigned long) on memory with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: memory is not defined"
FAIL Memory interface: memory must inherit property "buffer" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: memory is not defined"
FAIL Table interface: existence and properties of interface object assert_own_property: self does not have own property "Table" expected property "Table" missing
FAIL Table interface object length assert_own_property: self does not have own property "Table" expected property "Table" missing
FAIL Table interface object name assert_own_property: self does not have own property "Table" expected property "Table" missing
FAIL Table interface: existence and properties of interface prototype object assert_own_property: self does not have own property "Table" expected property "Table" missing
FAIL Table interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "Table" expected property "Table" missing
FAIL Table interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "Table" expected property "Table" missing
FAIL Table interface: operation grow(unsigned long) assert_own_property: self does not have own property "Table" expected property "Table" missing
PASS Unscopable handled correctly for grow(unsigned long) on Table
FAIL Table interface: operation get(unsigned long) assert_own_property: self does not have own property "Table" expected property "Table" missing
PASS Unscopable handled correctly for get(unsigned long) on Table
FAIL Table interface: operation set(unsigned long, Function) assert_own_property: self does not have own property "Table" expected property "Table" missing
PASS Unscopable handled correctly for set(unsigned long, Function) on Table
FAIL Table interface: attribute length assert_own_property: self does not have own property "Table" expected property "Table" missing
PASS Unscopable handled correctly for length property on Table
FAIL Global interface: existence and properties of interface object assert_own_property: self does not have own property "Global" expected property "Global" missing
FAIL Global interface object length assert_own_property: self does not have own property "Global" expected property "Global" missing
FAIL Global interface object name assert_own_property: self does not have own property "Global" expected property "Global" missing
FAIL Global interface: existence and properties of interface prototype object assert_own_property: self does not have own property "Global" expected property "Global" missing
FAIL Global interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "Global" expected property "Global" missing
FAIL Global interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "Global" expected property "Global" missing
FAIL Global interface: operation valueOf() assert_own_property: self does not have own property "Global" expected property "Global" missing
PASS Unscopable handled correctly for valueOf() on Global
FAIL Global interface: attribute value assert_own_property: self does not have own property "Global" expected property "Global" missing
PASS Unscopable handled correctly for value property on Global
FAIL CompileError interface: existence and properties of interface object assert_own_property: self does not have own property "CompileError" expected property "CompileError" missing
FAIL CompileError interface object length assert_own_property: self does not have own property "CompileError" expected property "CompileError" missing
FAIL CompileError interface object name assert_own_property: self does not have own property "CompileError" expected property "CompileError" missing
FAIL CompileError interface: existence and properties of interface prototype object assert_own_property: self does not have own property "CompileError" expected property "CompileError" missing
FAIL CompileError interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "CompileError" expected property "CompileError" missing
FAIL CompileError interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "CompileError" expected property "CompileError" missing
FAIL LinkError interface: existence and properties of interface object assert_own_property: self does not have own property "LinkError" expected property "LinkError" missing
FAIL LinkError interface object length assert_own_property: self does not have own property "LinkError" expected property "LinkError" missing
FAIL LinkError interface object name assert_own_property: self does not have own property "LinkError" expected property "LinkError" missing
FAIL LinkError interface: existence and properties of interface prototype object assert_own_property: self does not have own property "LinkError" expected property "LinkError" missing
FAIL LinkError interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "LinkError" expected property "LinkError" missing
FAIL LinkError interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "LinkError" expected property "LinkError" missing
FAIL RuntimeError interface: existence and properties of interface object assert_own_property: self does not have own property "RuntimeError" expected property "RuntimeError" missing
FAIL RuntimeError interface object length assert_own_property: self does not have own property "RuntimeError" expected property "RuntimeError" missing
FAIL RuntimeError interface object name assert_own_property: self does not have own property "RuntimeError" expected property "RuntimeError" missing
FAIL RuntimeError interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RuntimeError" expected property "RuntimeError" missing
FAIL RuntimeError interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RuntimeError" expected property "RuntimeError" missing
FAIL RuntimeError interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "RuntimeError" expected property "RuntimeError" missing
Harness: the test ran to completion.
